+++ xlators/performance/readdir-ahead/src/readdir-ahead.c
@@ -57,13 +57,13 @@ rda_fd_ctx *get_rda_fd_ctx(fd_t *fd, xlator_t *this)
 		/* ctx offset values initialized to 0 */
                 ctx->xattrs = NULL;
 
-		if (__fd_ctx_set(fd, this, (uint64_t) ctx) < 0) {
+		if (__fd_ctx_set(fd, this, (uint64_t)(uintptr_t) ctx) < 0) {
 			GF_FREE(ctx);
 			ctx = NULL;
 			goto out;
 		}
 	} else {
-		ctx = (struct rda_fd_ctx *) val;
+		ctx = (struct rda_fd_ctx *)(uintptr_t) val;
 	}
 out:
 	UNLOCK(&fd->lock);
@@ -548,7 +548,7 @@ rda_releasedir(xlator_t *this, fd_t *fd)
 	if (fd_ctx_del(fd, this, &val) < 0)
 		return -1;
 
-	ctx = (struct rda_fd_ctx *) val;
+	ctx = (struct rda_fd_ctx *)(uintptr_t) val;
 	if (!ctx)
 		return 0;
